Round-trip in-memory persistence through WASM (#72)
# Description
Routing the in-memory persistence through WASM first, before implementing a LocalStorage binding. Add `?w=1` to the URL to ignore whitespace changes.

We define a `Client` JS class that wraps the equivalent Rust object, intending for the interface to resemble that used by the JS SDK. We use a similar pattern to the Voodoo prototype here, where we hold Rust objects in memory in Rust-land and refer to them by ID's in JS-land. There could be some argument for moving the wrapper class directly into the JS SDK rather than implementing it here in the wasm binding library.

Dynamic traits such as the Persistence trait in this PR are not thread-safe by default. I am assuming that the entry-point into our WASM library is single-threaded, so I added the `Sync` and `Send` annotations to the `Persistence` trait to disable the compiler error about thread safety - feedback definitely welcome here.

# Test plan
`cargo test` in `xmtpv3/`
`npm run test` in `xmtpv3/bindings/wasm`