Add networking to XMTP client and expose it in FFI crate (#114)
Note that half of the changes are in `bindings_ffi/src/lib.rs`, which gets automatically hidden as a large diff.

This is the starting point for exposing our libxmtp client to the platform layer (iOS/Android). I added the networking trait from @neekolas's PR with some slight modifications and hooked up `xmtp_networking` to it.

There's still a lot of gaps to fill here but figured I should land this first so there's some initial structure:
1. Have not tested whether this actually works with `swift-bridge` at the platform layer. I will switch to Mozilla uniffi first and test the integration at that point.
2. Need to add proper error handling.
3. Have not exposed account creation to the platform layer yet.
4. Have not hooked up the API auth token.

I will work on (1) next - if anyone wants to pull any of the others feel free!