Cross platform builds in CI (#133)
This PR adds a GH workflow that builds across platforms, specifically the android platforms referenced in the cross_build.sh scripts, but can be extended to build for others. It seems like one of the next steps is to hook up version publishing in general for these packages/platforms, but this PR just focuses on getting the builds running in CI for continuous validation/testing and further iteration. The builds are still relatively slow in CI, taking about 15 minutes to complete, but not nearly as slow as when running locally on a mac via qemu.

https://github.com/bitcoindevkit/bdk-ffi/blob/master/.github/workflows/publish-android.yaml seems like a good example of the full build + publish workflow that we'd want to evolve this into.

Note that the `x86_64-linux-android` build is failing, with some context in https://github.com/xmtp/libxmtp/issues/134. I spent some time trying to figure this out, unsuccessfully, and ultimately decided to spin up an issue to track it separately and get a PR up for the builds that work.