Uniffi interface for xmtpv3 and Android example (#137)
This is a pretty barebones, scrappy initial version of running xmtpv3 end-to-end through an example Android app. The only methods exposed in this PR are creating the client and getting the account address from the client, which does a little bit under the hood, including networking calls.

I've tried to make the README.md in `bindings_ffi` as clear as possible, but please leave comments for anything I need to explain that I've missed.

The build is pretty slow. For now, dev workflow might be to just run `gen_kotlin.sh` to generate the bindings without doing any cross-compilation. This is relatively fast and can reveal any mistakes in the bindings interface. I'll add this to CI eventually.

Coming up next:
- Mutable function calls
- Pass wallet down from app
- Error handling strategy
- Run gen_kotlin.sh from CI (synchronously)
- Cross-compile from CI (asynchronously) - and make it faster ideally
- Run android integration test from CI (asynchronously)

![Screen Shot 2023-06-09 at 3 43 47 PM](https://github.com/xmtp/libxmtp/assets/696206/2dffe71e-3054-4c4c-bdf2-2a564d8f73b1)