Remove TECHNICAL_OVERVIEW.md (#499)
This document is extremely out-of-date and currently creates more confusion than clarity. We can re-use pieces of it when we create a technical overview for the MLS work, but in the meantime `xmtp_mls/README.md` and `xmtp_mls/IDENTITY.md` can give a partial overview of the technical design.