Teach example/CLI how to talk to V4/Replication Node (#1214)
This is the first (large) step in the right direction. By no means is the migration to decentralization complete.

This teaches the CLI to do basic operations via xtmpd such as:
- registering
- creating groups
- adding members to groups
- messaging

One has to simply provide the `--testnet` flag to the CLI and everything will work against the new API.

This refactors the Error class into its own file and adds an Internal error.

Also adds a new command to the CLI for debugging.