Ability to clear network stats (#2085)
Because we cache the GRPC client on mobile the network stats can get
very large quickly and it's hard to tell what the actual network calls
were that were being made. This give us better control to be able to
clear the stats run an action and see exactly what is happening.