Commit log flag (#2176)
Disables the commit log for now, so that we don't have incomplete data
on production clients when the feature is eventually released.

This PR enables the feature flag for tests only. There are pros and cons
of this approach.
- Pro: Allows us to write unit tests against the commit log as we
develop the feature
- Con: If we regress something when the flag is turned off, we won't
catch it via unit test

To mitigate the con, I think we can just be careful that changes outside
of the flag don't depend on logic inside the flag. We can also run the
unit test with the flag off as a sanity check before landing a PR.

In the long-term, there's probably work we can do to make an INIT_ONCE
config object that holds the flags, and allow specific tests to override
it, rather than enabling the flag across all unit tests. There's various
other places in the codebase that could use this too.