Add welcomes to local commit log (#2180)
When Welcomes have a cursor pointing to where in the group the newly
added installation should sync from (#2088), we should also be able to
use this as the `commit_sequence_id` in the local/remote commit log.
This would allow a newly added installation to immediately check if they
were added into a fork, tackling the viral nature of forks today.

Because we don't have this information yet, we can instead store it with
a sequence_id of 0 for local debugging purposes for now. Anything with a
sequence_id of 0 won't be uploaded to the remote commit log, and is
effectively ignored for the purposes of fork detection.