Improved commit log error handling (#2185)
Changes include:
1. After joining a group from a Welcome message, do not add errored
commit log entries for commits sent before the installation joined the
group. When we get an epoch error, we do this by checking if the last
entry was a Welcome, or if there is no previous entry. Worth noting:
  - Even if we implement logic to add a cursor to Welcomes and only sync
from that cursor onwards, we cannot rely on old clients populating this
cursor when they send us a Welcome
  - When we get an epoch error, we can check if the last commit log entry
is a Welcome to determine whether to ignore it, but we can't rely on
that entry being present - the client may have received a Welcome on an
older version, and not have written that entry to the log.
2. Populate the `epoch_number` and `epoch_authenticator` for error
commit logs from the actual group state, rather than inferring it from
the most recent entry

We also need to be careful about how we handle Welcomes that are readds
(for example, as part of automated fork recovery, or after leaving and
re-joining the group). I chose not to clear the local commit log on a
readd, so that we have a clearer audit trail for debugging. Instead,
when comparing the local commit log with the remote commit log, we can
stop comparing when we reach the most recent Welcome. I've updated the
XIP to describe this.

Next PR will add more unit tests for errors, and have some light
refactoring.