Add `should_publish_commit_log` to groups table (#2197)
See tracking issue: https://github.com/xmtp/libxmtp/issues/2102

The criteria for whether we should publish our local commit log to the remote commit log is whether we are a super admin for the group.

We need to store this in the db so we dont have to check the mls group state for each group every time we publish to the remote commit log. 

The security model of the remote commit log does not depend on super admin status being verified on the publish side, so this is sort of a best effort to accurately assess that only super admins publish to the remote commit log. If a non super admin publishes to the remote commit log, we verify downstream in the fork resolution process that we will not accept a re-add from a non super-admin. See https://community.xmtp.org/t/xip-68-draft-automated-fork-recovery/951