Create builder for intents  (#2250)
The builder makes it easier to write unit tests, as well as makes the
code when working with bulk groups clearer.

**Reasoning:**
Q: If we are to batch intents, how does the function fit into libxmtp

`queue_intents` is implemented on the group. having a method on group
accept multiple other groups where the intent would be queued seems
confusing (why does one group get "priority"?). + the ergonomics of
randomly choosing a group to queue an intent on many other groups. The
intent is the same for all groups, so it shouldn't rely on getting any
singular group.

In that case, we could put the batch queue on Client. But, now we have
separated logic/places for queuing intents, which again seems like it
would be confusing and might lead to bugs. There's also many other
functions implemented on client, which can make encapsulating
functionality tricky


Making intents a first-class citizen improves ergonomics, and also lets
us write better tests for the queueing logic itself, while keeping the
implementation details private