Fix app_version accessor (node bindings) (#2293)
### Fix app_version accessor in node bindings by changing Client struct
field visibility and replacing property access with method call
Changes the `app_version` field in the `Client` struct from public to
private in
[bindings_node/src/client.rs](https://github.com/xmtp/libxmtp/pull/2293/files#diff-98dd73d0e8b1dce0a25de3ddbda3a0122baec850f3b6ccbdca525c3370f80f1d)
and updates test cases in
[bindings_node/test/Client.test.ts](https://github.com/xmtp/libxmtp/pull/2293/files#diff-32a11487468143d5cfc69e5457f7bfab2f05c4a486fef691bd8796384326c95e)
to use `client.appVersion()` method call instead of direct property
access, with the method returning an empty string instead of null when
no app version is set.

#### 📍Where to Start
Start with the `Client` struct definition in
[bindings_node/src/client.rs](https://github.com/xmtp/libxmtp/pull/2293/files#diff-98dd73d0e8b1dce0a25de3ddbda3a0122baec850f3b6ccbdca525c3370f80f1d)
to see the visibility change of the `app_version` field.

----

_[Macroscope](https://app.macroscope.com) summarized a681c70._